# Conflicts:
#	CMakeLists.txt
#	Common/CMakeLists.txt
#	Common/ac/game_version.h
#	Common/ac/gamestructdefines.h
#	Common/ac/spritecache.cpp
#	Common/core/def_version.h
#	Common/font/fonts.cpp
#	Common/font/fonts.h
#	Common/game/main_game_file.cpp
#	Common/gui/guibutton.cpp
#	Common/gui/guilistbox.cpp
#	Common/gui/guimain.cpp
#	Editor/AGS.Editor/AGSEditor.cs
#	Editor/AGS.Editor/DataFileWriter.cs
#	Editor/AGS.Editor/Tasks.cs
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Native/NativeMethods.cpp
#	Editor/AGS.Types/AGS.Types.csproj
#	Editor/AGS.Types/Enums/FontAutoOutlineStyle.cs
#	Editor/AGS.Types/Font.cs
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Editor/Native/sprcache_agsnative.cpp
#	Engine/ac/character.cpp
#	Engine/ac/display.cpp
#	Engine/ac/draw.cpp
#	Engine/ac/draw.h
#	Engine/ac/drawingsurface.cpp
#	Engine/ac/gui.cpp
#	Engine/ac/gui.h
#	Engine/ac/overlay.cpp
#	Engine/ac/overlay.h
#	Engine/ac/runtime_defines.h
#	Engine/ac/screenoverlay.cpp
#	Engine/ac/screenoverlay.h
#	Engine/ac/statobj/staticgame.h
#	Engine/game/game_init.cpp
#	Engine/game/savegame_components.cpp
#	Engine/game/savegame_v321.cpp
#	Engine/gfx/ali3dogl.h
#	Engine/gfx/ali3dsw.h
#	Engine/gfx/blender.cpp
#	Engine/gfx/blender.h
#	Engine/gfx/ddb.h
#	Engine/gfx/gfxdriverbase.h
#	Engine/main/engine.cpp
#	Engine/platform/android/acpland.cpp
#	Engine/platform/base/agsplatformdriver.h
#	Engine/platform/windows/gfx/ali3dd3d.h
#	version.json

+++ Compiler/preproc/preprocessor.cpp
@@ -152,7 +152,7 @@ namespace Preprocessor {
     }
 
     String Preprocessor::GetNextWord(String &text, bool trimText, bool includeDots) {
-        int i = 0;
+        size_t i = 0;
         while ((i < text.GetLength()) &&
                (is_alphanum(text[i]) ||
                 (includeDots && (text[i] == '.')))
@@ -181,7 +181,7 @@ namespace Preprocessor {
         }
 
         StringBuilder output = StringBuilder(text.GetLength());
-        for (int i = 0; i < text.GetLength(); i++)
+        for (size_t i = 0; i < text.GetLength(); i++)
         {
             if (!_inMultiLineComment)
             {
